+++ packData/unpackData.mjs
@@ -31,7 +31,8 @@ let itemPacks = [
     'cpr-monster-features-2024',
     'cpr-3rd-party-spells-2024',
     'cpr-species-features-2024',
-    'cpr-other-features'
+    'cpr-other-features',
+    'cpr-3rd-party-feats-2024'
 ];
 let actorPacks = [
     'cpr-summons',